fixes based on comments
