+++ test/e2e/utils/helper.go
@@ -10,7 +10,6 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/google/go-cmp/cmp"
 	// Lint check prohibits non "_test" ending files to have dot imports for ginkgo / gomega.
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
@@ -209,14 +208,4 @@ func GenerateCRDObjectFromFile(cluster framework.Cluster, fs embed.FS, filepath 
 	gomega.Expect(err).Should(gomega.Succeed(), "CRD data was not mapped in the restMapper")
 
 	return obj, gvk, mapping.Resource
-}
-
-func IsKeyMetadata(p cmp.Path) bool {
-	step, ok := p[len(p)-1].(cmp.MapIndex)
-	return ok && step.Key().String() == "metadata"
-}
-
-func IsKeyNotName(p cmp.Path) bool {
-	step, ok := p[len(p)-1].(cmp.MapIndex)
-	return ok && step.Key().String() != "name"
 }
